requirements.txt
GitPython>=3.1.40
colorama>=0.4.6
PyYAML>=6.0.1
